break out query commands
