+++ packages/driver/src/cy/commands/querying/index.ts
@@ -1,0 +1,15 @@
+import * as Contains from './contains'
+import * as Focused from './focused'
+import * as Get from './get'
+import * as Root from './root'
+import * as Shadow from './shadow'
+import * as Within from './within'
+
+export {
+  Contains,
+  Focused,
+  Get,
+  Root,
+  Shadow,
+  Within,
+}
